[llvm] The function specializer should use function cloning instead of relying on the inliner to perform specialization
The function specializer was relying on the inliner to kick in and produce a specialized body of a function by substituting constant arguments. This works for simple cases, but does not allow for performing any specialization before inlining, which is required in some cases.

This commit uses LLVM's cloning utilities to create the specialized body and makes the function specializer self-contained and not dependent on the LLVM's optimizer passes.