[ir] Represent Glow IR instructions lists using LLVM's intrusive lists
This change solves a number of compilation performance issues, because instructions can be found in the list or removed from the list very quickly.

The change touches many of the loops iterating over the instruction lists. In most cases, the loop bodies are interested in the pointers to instructions, but the default LLVM's intrusive list iterators return references to the instructions. Using such iterators directly would require updating the bodies of most such loops. Therefore, a convenience method called ForElementPtrIterator is introduced, which provides iterators returning pointers to instructions and allows to keep loop bodies as is.

While looking at the loops iterating over the Glow IR instructions, also take the chance to tighten their const correctness. This includes tightening the const correctness of some APIs used by those loops as well.