Make sure that use-lists of Nodes only contain uses by other Nodes
Until now, NodeValue was used to refer to the results of Nodes. But unfortunately, every NodeValue would add itself to the use-list of a Node it references, which is rather expensive. In many cases, NodeValue objects are created as temporary objects and it makes no sense to consider them to be uses.

We want to count only references from Nodes to the results of other Nodes as uses. To achieve that this commit introduces a new type NodeValueHolder. It is used instead of NodeValue inside Nodes to refer to the results of other nodes. All NodeValueHolders add themselves as uses to the use lists. And NodeValue is changed in such a way that it does not register itself in the use-lists anymore.

To prove that there are no unexpected addUse calls in the whole project, I added a "backtrace" method that detects all bad attempts to call addUse by analyzing the call stacks of addUse calls.